refactor: move settings into seperate folder and destrcut props inline the function
