+++ packages/form-builder/src/blocks/fields/index.js
@@ -20,9 +20,5 @@ const fieldBlocks = [
     amount,
 ];
 
-const fieldBlockNames = fieldBlocks.map(field => field.name);
 
 export default fieldBlocks;
-export {
-    fieldBlockNames,
-};
