+++ packages/form-builder/src/blocks/fields/settings/index.jsx
@@ -1,0 +1,44 @@
+import {__} from '@wordpress/i18n';
+import Edit from './Edit';
+
+const defaultSettings = {
+    title: __('Field', 'custom-block-editor'),
+
+    supports: {
+        html: false, // Removes support for an HTML mode.
+        multiple: true,
+    },
+
+    attributes: {
+        fieldName: {
+            type: 'string',
+            source: 'attribute',
+        },
+        label: {
+            type: 'string',
+            source: 'attribute',
+            default: __('Text Field', 'give'),
+        },
+        placeholder: {
+            type: 'string',
+            source: 'attribute',
+            default: '',
+        },
+        isRequired: {
+            type: 'boolean',
+            source: 'attribute',
+            default: false,
+        },
+        options: {
+            type: 'array',
+        },
+    },
+
+    edit: Edit,
+
+    save: function () {
+        return null; // Save as attributes - not rendered HTML.
+    },
+};
+
+export default defaultSettings;
